+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Stripping away lazy resolver
 - FIX: Resolving to duplication that resolved to UNDEFINED should raise an exception
 - Add performance test (pyhocon comparison)
+- FIX: Resolve duplication with a concatenation in between that resolves to UNDEFINED
 
 ## 0.6.2
 - Simplify lazy deduplication